리뷰준비

+++ base/src/main/java/smart/util/FileProviderHelper.kt
@@ -10,20 +10,41 @@ import java.io.IOException
 import java.text.SimpleDateFormat
 import java.util.*
 
+/**
+ * ```
+ *     <uses-permission
+ *         android:name="android.permission.WRITE_EXTERNAL_STORAGE"
+ *         android:maxSdkVersion="18" />
+ *     <application>
+ *         <provider
+ *             android:name="androidx.core.content.FileProvider"
+ *             android:authorities="${applicationId}.provider"
+ *             android:enabled="true"
+ *             android:grantUriPermissions="true"
+ *             android:exported="false">
+ *             <meta-data
+ *                 android:name="android.support.FILE_PROVIDER_PATHS"
+ *                 android:resource="@xml/provider_paths" />
+ *         </provider>
+ *
+ *         <activity
+ *             android:name="smart.util.GalleryLoader"
+ *             android:theme="@style/GalleryLoaderTheme" />
+ *
+ *     </application>
+ * ```
+ */
+
 object FileProviderHelper {
     private const val PROVIDER = ".provider"
     private val GLOADER_FOLDER = "gloader" to "galleryloader/"
     private val GTEMP_FOLDER = "gtemp" to "galleryloader/temp/"
     private fun getRootFolder(context: Context): File {
-        var root = context.getExternalFilesDir(null)
-        if (root == null)
-            root = context.filesDir
-        return mkdirs(root!!)
+        return mkdirs(context.getExternalFilesDir(null)!!)
     }
 
-    private fun getFolder(context: Context, sub_folder: String): File {
-        val folder = File(getRootFolder(context), sub_folder)
-        return mkdirs(folder)
+    private fun createFolder(context: Context, sub_folder: String): File {
+        return mkdirs(File(getRootFolder(context), sub_folder))
     }
 
     private fun mkdirs(dir: File): File {
@@ -37,79 +58,51 @@ object FileProviderHelper {
     }
 
     private fun toFile(context: Context, uri: Uri?): File? {
-        uri?.run {
-            if (scheme == "content" && authority == context.packageName + PROVIDER && !path.isNullOrBlank()) {
-                for (pair in arrayOf(GTEMP_FOLDER, GLOADER_FOLDER)) {
-                    val (name, folder) = pair
-                    if (name in path!!) {
-                        val pathname = path!!.replaceFirst(name, folder)
-                        Log.e(path)
-                        Log.e("=>")
-                        Log.e(pathname)
-                        val file = File(getRootFolder(context), pathname)
-                        Log.e(uri)
-                        Log.e("=>")
-                        Log.e(file)
-                        return file
-                    }
-                }
+        uri ?: return null
+        if (uri.scheme != "content") return null
+        if (uri.authority != context.packageName + PROVIDER) return null
+        if (uri.path.isNullOrBlank()) return null
+
+        for (pair in arrayOf(GTEMP_FOLDER, GLOADER_FOLDER)) {
+            val (name, folder) = pair
+            if (name in uri.path!!) {
+                val pathname = uri.path!!.replaceFirst(name, folder)
+                val file = File(getRootFolder(context), pathname)
+                Log.e(file to uri)
+                return file
             }
         }
         return null
     }
 
-    //    private fun getFile(context: Context, filename: String): File {
-//        val file = File(getFolder(context, GLOADER_FOLDER.second), filename)
-//        Log.e(file)
-//        return file
-//    }
-    private fun getFile(context: Context, prefix: String = "", suffix: String = ".jpg"): File {
-        val timeStamp: String = SimpleDateFormat("yyyyMMdd_HHmmss", Locale.getDefault()).format(Date())
-        val file = File.createTempFile("${prefix}_${timeStamp}_", suffix, getFolder(context, GLOADER_FOLDER.second))
-        Log.e(file)
-        return file
-    }
-
-    //    fun getUri(context: Context, filename: String): Uri {
-//        val file = getFile(context, filename)
-//        return getUri(context, file)
-//    }
-    private fun getUri(context: Context, file: File): Uri {
+    private fun toUri(context: Context, file: File): Uri {
         val authority = context.packageName + PROVIDER
         val uri = FileProvider.getUriForFile(context, authority, file)
-        Log.e(file)
-        Log.e("=>")
-        Log.e(uri)
+        Log.e(uri to file)
         return uri
     }
 
     @Throws(IOException::class)
-    fun getTempFile(context: Context, prefix: String, suffix: String?): File {
+    private fun createTempFile(context: Context, prefix: String = "", suffix: String? = ".jpg"): File {
         val timeStamp: String = SimpleDateFormat("yyyyMMdd_HHmmss", Locale.getDefault()).format(Date())
-        return File.createTempFile("${prefix}_${timeStamp}_", suffix, getFolder(context, GTEMP_FOLDER.second))
+        return File.createTempFile("${prefix}_${timeStamp}_", suffix, createFolder(context, GLOADER_FOLDER.second))
     }
 
     @Throws(IOException::class)
-    fun getTempUri(context: Context, prefix: String, suffix: String?): Uri {
+    fun createTempUri(context: Context, prefix: String, suffix: String?): Uri {
         val authority = context.packageName + PROVIDER
-        val file = getTempFile(context, prefix, suffix)
+        val file = createTempFile(context, prefix, suffix)
         val uri = FileProvider.getUriForFile(context, authority, file)
-        Log.e(file)
-        Log.e("=>")
-        Log.e(uri)
+        Log.e(uri to file)
         return uri
     }
 
-    //    fun deleteFile(context: Context, uri: Uri?): Boolean {
-//        Log.e(uri)
-//        toFile(context, uri)?.run {
-//            return exists() && delete()
-//        }
-//        return false
-//    }
-    fun deleteTemp(context: Context) {
-        val source = getFolder(context, GTEMP_FOLDER.second)
-        val target = File(getRootFolder(context), System.nanoTime().toString())
+    fun deleteTempFolder(context: Context) {
+        val source = File(getRootFolder(context), GTEMP_FOLDER.second)
+        if (!source.exists()) return
+//        val source = createFolder(context, GTEMP_FOLDER.second)
+        val target = File.createTempFile("delete_", "", getRootFolder(context))
+//        createTempFile(getRootFolder(context), System.nanoTime().toString())
 
         if (source.renameTo(target)) {
             deleteRecursive(target)
@@ -128,10 +121,10 @@ object FileProviderHelper {
     fun moveForResult(context: Context, uri: Uri?): Uri? {
         return try {
             val source = toFile(context, uri)
-            val target = getFile(context, "result", ".jpg")
+            val target = createTempFile(context, "result", ".jpg")
             val result = source?.renameTo(target)
             if (result!!)
-                getUri(context, target)
+                toUri(context, target)
             else
                 uri
         } catch (e: Exception) {
@@ -140,7 +133,7 @@ object FileProviderHelper {
     }
 
     fun copyForCrop(context: Context, source: Uri?): Uri {
-        val file = FileProviderHelper.getTempFile(context, "crop", ".jpeg")
+        val file = FileProviderHelper.createTempFile(context, "crop", ".jpeg")
         source?.also {
             val inputStream = context.contentResolver.openInputStream(it)
             val outputStream = FileOutputStream(file)
@@ -148,6 +141,6 @@ object FileProviderHelper {
             inputStream?.close()
             outputStream.close()
         }
-        return getUri(context, file)
+        return toUri(context, file)
     }
 }
